Made adjustments: added navbar, created new html for calculator.q

+++ scripts/calculator.js
@@ -10,16 +10,13 @@ function init() {
     number1Field = document.getElementById("number1Field");
     number2Field = document.getElementById("number2Field");
 
-    // let addButton = document.getElementById("addButton");
+    // document.getElementById("addButton").onclick = onAddButtonClicked;
     addButton.onclick = onAddButtonClicked;
-
-    // let subtractButton = document.getElementById("subtractButton");
+    
     subtractButton.onclick = onSubtractButtonClicked;
-
-    // let multiplyButton = document.getElementById("multiplyButton");
+   
     multiplyButton.onclick = onMultiplyButtonClicked;
 
-    // let divideButton = document.getElementById("divideButton");
     divideButton.onclick = onDivideButtonClicked;
 
     clearButton.onclick = onClearButtonClicked;
@@ -27,22 +24,22 @@ function init() {
 }
 
 function onAddButtonClicked() {
-    answerField = parseInt(number1Field.value) + parseInt(number2Field.value); 
+    answerField = parseFloat(number1Field.value) + parseFloat(number2Field.value); 
     document.getElementById("answerField").value = answerField;
 }
 
 function onSubtractButtonClicked() {
-    answerField = parseInt(number1Field.value) - parseInt(number2Field.value); 
+    answerField = parseFloat(number1Field.value) - parseFloat(number2Field.value); 
     document.getElementById("answerField").value = answerField;
 }
 
 function onMultiplyButtonClicked() {
-    answerField = parseInt(number1Field.value) * parseInt(number2Field.value); 
+    answerField = parseFloat(number1Field.value) * parseFloat(number2Field.value); 
     document.getElementById("answerField").value = answerField;
 }
 
 function onDivideButtonClicked() {
-    answerField = parseInt(number1Field.value) / parseInt(number2Field.value); 
+    answerField = parseFloat(number1Field.value) / parseFloat(number2Field.value); 
     document.getElementById("answerField").value = answerField.toFixed(2);
 }
